// frontend/src/models/Bug.ts
export interface Bug {
  id: string
  name: string
  iconImage: string
  critterpediaImage: string
  furnitureImage: string
  sellPrice: number
  whereHow: string
  weather: string
  totalCatchesToUnlock: number
  spawnRates: string
  nhJan: string
  nhFeb: string
  nhMar: string
  nhApr: string
  nhMay: string
  nhJun: string
  nhJul: string
  nhAug: string
  nhSep: string
  nhOct: string
  nhNov: string
  nhDec: string
  shJan: string
  shFeb: string
  shMar: string
  shApr: string
  shMay: string
  shJun: string
  shJul: string
  shAug: string
  shSep: string
  shOct: string
  shNov: string
  shDec: string
  size: string
  surface: string
  description: string
  catchPhrase: string
  hhaBasePoints: number
  hhaCategory: string
  color1: string
  color2: string
  donated: boolean
  critterpediaOrder: number
  flickSellPrice: number
  rarity: string
}